Refactor: Reorganize codebase with domain-specific type and utility separation
