+++ src/agent/types/ResultTypes.ts
