+++ src/frontend/files/fileLister.ts
@@ -10,7 +10,7 @@ import * as logger from '@logger/logUtils';
 // Local imports - utilities
 import { getConfig, watchConfig } from '@utils/config';
 import { WorkspaceFS } from '@utils/files';
-import { getIncludedExtensions, FileType } from '@utils/fileTypeUtils';
+import { getIncludedExtensions, FileType } from '@common/files/fileTypeUtils';
 import { getFilesInDirectory, getFilesRecursively } from './listing';
 
 const CHANNEL = 'FileLister';
