+++ src/logger/LogTypes.ts
@@ -3,8 +3,8 @@
  */
 
 // Local imports
-import type { TokenUsageStats } from '../types/UsageTypes';
-import type { TaskGroupId, LogMessageId } from '../types/EntityTypes';
+import type { TokenUsageStats } from '@agent/types/UsageTypes';
+import type { TaskGroupId, LogMessageId } from './types/EntityTypes';
 
 export interface TaskGroup {
   /** Unique identifier for the group */
