+++ src/logger/types/EntityTypes.ts
