+++ src/utils/system/execUtils.ts
@@ -6,7 +6,7 @@ import * as logger from '@logger/logUtils';
 
 // Local imports - utilities
 import { WorkspaceFS } from '@utils/files';
-import { ExecResult } from '@/types/ResultTypes';
+import { ExecResult } from '@agent/types/ResultTypes';
 import { extendEnvPath } from '@utils/system/platformPaths';
 
 const CHANNEL = 'execUtils';
